requirements.txt
apiclient==1.0.3
cffi==1.7.0
cryptography==1.4
decorators==0.1
endpoints==1.1.16
enum34==1.1.6
google-api-python-client==1.5.1
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
ndg-httpsclient==0.4.1
oauth2client==2.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyfcm==0.0.4
pyOpenSSL==16.0.0
requests==2.10.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
sparkpost==1.2.0
uritemplate==0.6
urllib3==1.16
